irc: remove unneeded server disconnect when server buffer is closed and server is already disconnected
This commit fixes a problem with the script autoconnect.py: the script hooks
signal "irc_server_disconnected", and this signal was sent 2 times for each
server on /quit: one time when servers are disconnected, and one time when
servers are destroyed (because buffer is closed, and then server disconnected
again). The script forces save of irc.conf on each disconnection, so some
servers were lost in irc.conf, when some servers have already been destroyed.