irc: fix prefix color for nick when the prefix is not in irc.color.nick_prefixes: use default color (key "*")
Problem was happening on a server which has "PREFIX=(Yqaohv)!~&@%+".
Users with prefix "!" were displayed as lightred (color for "~") instead of
lightblue (default key "*").

When a prefix was not found, WeeChat was looping on other prefixes (in
order). Now if color is not found, WeeChat uses immediately the fallback color.