core: revert check of layout when switching to a buffer using value "auto" in buffer_set
This reverts commits 14cf7bd20fb3f00b18343464c175a0c4e3556149 and
262dff4b87ebf14a11fe83899fe168bc5a72ffe6.

The initial fix was for a problem with irc autojoin and layout: the
current window has not appropriate buffer (as defined in layout) if it
is not the latest channel in "autojoin" option.

So the way to fix this problem is to set the value of option
irc.look.buffer_switch_autojoin to off.