core: fix random crash when closing a buffer
The problem happened because we used a pointer to a
"struct t_gui_buffer_visited" for the switch to another buffer,
when the buffer is closed. This is executed in all windows displaying
the buffer, but on each switch to buffer, the visited buffers are
updated and therefore the address can change. The pointer becomes
invalid, and WeeChat still uses it on next windows for the buffer
switch.

It happened rarely because the visited buffer is freed and allocated
immediately after, so the address is often the same in memory.

Thanks to silverd for the tests on OS X to track the problem.