ruby: add detection and fix compilation with Ruby 2.0 (patch #8209)
Now the search for Ruby is first performed with pkg-config and includes
detection of Ruby 2.0.
If not found, the old code for detection is used (for old distros or
old Ruby versions).

The specific test on Ruby 1.9.0 (to disable this version) has been
removed (1.9.0 was a dev/unstable version from 2007, quite old now).