core: fix "/window scroll_bottom" on a buffer with free content
The bug was introduced by commit 8fac1eea40d6306c27bd854f8ae7bd0c35632fc9
which purpose was to not scroll to the end of buffer with "/window scroll -N"
when the top of buffer is displayed.

This commit is adding the syntax "--N" to force a scroll from the end (only
used by "/window scroll_bottom").