irc: fix truncated read on socket with SSL (bug #41558)
If there are still data on socket with SSL
(if gnutls_record_check_pending(session) returns > 0), then call
gnutls_record_recv() again to read all available data.

The problem was that some data remained in the gnutls buffers and the recv
callback was not called any more (nothing available on raw socket).