core: fix crash on /reload weechat
The problem was introduced in commit eef3b570752b974b87f5b189c89813f400f55a6e
which was storing split commands inside keys themselves (for optimization).

This causes a problem when weechat.conf is reloaded, because all keys are
then deleted and created again, which removes the array of commands which
is currently used. And then it leads to a crash of WeeChat.

This commit reverts this behavior and the key commands are split when they
are executed.