core: add non-active merged buffers with activity in hotlist (if another merged buffer is zoomed) (task #12845)
There was a bug with "num_displayed" in buffers when they are merged and
that one buffer is zoomed: now the num_displayed is > 0 for the active buffer,
and it is set to 0 for all merged non-active buffers (only in case of zoom).

A variable "zoomed" has been added in buffers to know if a buffer with this
number is zoomed or not (possibly another buffer).