.travis.yml: install msgcheck directly from git
and also upgrade it and pylint if they happen to be installed.

As it is @FlashCode 's repository, it's probably as stable as WeeChat
git and it might contain changes that they didn't remember to send to
git, so I think that this is wanted change.

If this isn't, this can be closed as this is just a suggestion and pull
request.