En/decode must be done on the whole irc message.
Message decoding must be done on the whole message.  Protocol commands
and control delimiters are all ascii subset of utf-8 and every
data from the server (like nickname, user ident, and server name) are
subject to charset encodings, so no need to focus only on text data
part.

Also, it has been detected that server data on output is first encoded
as a message and then charset converted.  This must happen in reverse
direction (first decoded, then parsed) on input, and message fields are
first extracted from message (to dynamic memory) and then the message
is converted (so message already extracted parameters are not charset
decoded)  This is wrong.  Message decoding is being moved to first place
(even before raw printing)