Move the decoding code to the beginning
The decoding code has ben moved to the beginning of
the input processing code.  In this way, we first
decode the message and then extract the info, not as
was being done before, that we extracted the info to
dynamically allocated strings and then did the decoding
of the original message (this lead to confusion on nicks
channels with non ascii characters)

This tried to be solved by trying to decode/encode only
parts of the message which is incorrect, so we have undone
that modifications.