Python plugin shared libraries missing
When linking against a python with shared libraries, this script ONLY asked for the LD flags, not any additional libraries.  This could result in a condition where required libraries (such as libutil on Centos 6.7 against Python 2.7) are not loaded and the load of the plugin thus fails.  This change asks the python being linked against which libraries it was linked using, and then ensures those are also linked against by the python plugin for weechat.