CMake: Warn about missing dependency
When an feature has been enabled but the dependency hasn't been
fulfilled, it now prints an warning. Please note that the minimum
version is CMake 2.4, but the message(WARNING ...) has been introduced
to CMake 2.8.

Also, there's the option STRICT, that lets the generation fail on such
error. This is OFF by default to not to break existing builders.