irc: implement delegated SASL ECDSA NIST256P mechanism
In order to minimize code changes, this mechanism assumes that
SASL_KEY option holds a path of the external program to be used
as an authentication agent. During authentication process, mechanism
implementation executes this agen, passing username and challenge data
to it. The output of the agent is passed to the server as challenge
response.