gui: attempt to auto-reset color pairs even when full
Previously if color pairs fill up because all pairs are rendered at once
and auto-reset therefore fails, weechat will never try to auto-reset
again and manual reset is necessary.

This attempts to auto-reset even a full table, in case now less pairs
are visible and the reset will be successful.