Annotate public symbols and hide the rest
Currently we export all the symbols from both the weechat executable and
each of the plugins.

Introduce WEECHAT_PUBLIC macro annotating the respective symbols and
hide everything else.

As a result, we shave around 10% of the size of each binary. Savings
should be even more if LTO is used.

Note: since cmake is the only officially supported build system, I did
not bother with autotools.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>