relay: UNIX socket support (implement #733)
This adds support for relay servers listening on UNIX domain sockets.
They can be created using the /relay addpath command and specifying the
un protocol option. 'path' is now used instead of 'port' in messages,
which contains either a string representation of the port or the path to
a socket; port is negative when a UNIX socket is used. Path based relays
are stored in their own config section, with existing callbacks reused
as much as possible. Documentation has been updated as well.