cmake: vendor in a new version of FindPkgConfig
The current one is anciently ancient, and dates back to commit
4d2925ef1c3dd5df288be55557358b60cbb0926e which vendored this "for old
versions of cmake". Well, currently it just stops using new versions of
FindPkgConfig, so we're stuck on the 2006 version from cmake 2.5.0.

Instead of deleting it entirely (the minimum version of cmake is
currently 3.0) make this vendored file continue to be useful by using it
to vendor in the latest version of FindPkgConfig from cmake 3.16.0-rc3
with a bunch of useful improvements.