plugins: irc: support server certificate public key fingerprints
This allows users to pin known public key fingerprints instead of
certificate fingerprints, which means that the fingerprints will
not change when the server's certificate is renewed (unless that
is accompanied by a change of server private key). With the
advent of short-lifetime certificates from the likes of Let's
Encrypt, this is a very useful feature.

This is not any more insecure than a certificate fingerprint;
server certificates only exist to securely bind a server name to
a cryptographic identity (its public key), so that you have some
degree of confidence that you are talking to the correct server,
and not an impostor (MITM). If you're establishing its crypto-
graphic identity up-front (by pinning a known-good fingerprint of
it with a cryptographically-secure digest algorithm), then the
certificate is, by definition, not needed. In fact, this is
slightly more secure than a certificate fingerprint, because
certificates are a lot more malleable; the ability to pad and
inject random data to aid in e.g. a second preimage attack is
much more easily achieved with an X.509 certificate than it is
with a public key (because you can generate a valid private key,
compute its corresponding public key, and then mess with other
fields in the certificate to affect the digest instead). Also,
any attack against Hash(publickey) implies an attack against
Hash(certificate), which would break the CA PKI model anyway.

Specifically, this fingerprints the entire SubjectPublicKeyInfo
structure from the certificate, which includes not just the key
itself, but also other meta-information about it (like its
algorithm identifier and any associated parameters). This is the
same public key fingerprinting approach taken by e.g. the HSTS
and DANE TLSA standards, and also some IRC servers (such as
Charybdis IRCd).

While working on this commit, I noticed a few oddities and/or
deficiencies in the existing fingerprinting code, which this
commit also corrects and/or improves:

- Several for() loops used an unrelated variable to iterate over
  arrays of fingerprint variables (for setting their initial
  values or freeing them after).

- When displaying that the server's fingerprint has matched the
  client's configuration, it did not specify which fingerprint
  was received from the server for matching. This is relevant in
  the case of a comma-separated list of fingerprints.

- The code checked if a variable was non-NULL before calling
  free(3) on it, but free(3) is well-defined to do nothing if
  passed NULL.

cf. weechat/weechat#879

Signed-off-by: Aaron Jones <aaron@alphachat.net>