guile: fix printing of output to buffer (issue #1098)
The support for Guile 2.2 did not implement the custom port correctly,
and did not configure it to be the default output port. This caused
output from Guile to be printed straight to the TTY, breaking the
WeeChat interface.

Note that the port is unbuffered, so that code like

    (display "test")

immediately results in output, without an explicit call to force-output.