core: add weechat.look.hotlist_update_on_buffer_switch and weechat.look.read_marker_update_on_buffer_switch
Both default to "on" (no change to previous behaviour).

If hotlist_update_on_buffer_switch is "off", then when switching away from a buffer, the hotlist is not cleared.

If read_marker_update_on_buffer_switch is "off", then when switching away from a buffer, the read marker is not set to the end of the buffer.

This allows users to choose to have fully manual control over when hotlist counts and buffer read markers are reset.

Fixes #992.