cmake: require 3.12, resolve circular dependency
In order to resolve the circular dependency, we need to annotate the
respective static libraries as object "libraries."

This requires cmake 3.12, where Debian 10 (old old stable) and Ubuntu
20.04 have 3.13 and 3.16 respectively.

An alternative is to manually track and use the respective source files.
Since all platforms of interest meet the cmake requirement, we can opt
for the simpler approach.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>