core: fix bar_item_line for wrapped/new lines with horizontal filling
When the bar item filling is horizontal and the text of an item is
either wrapped or contains a \r so it's multiple lines and you focus on
a line that isn't the first, the bar_item_line value did not represent
the line you focused on in the bar item.

Instead, it represented which line the character you focused on
originally was in the string used to create the bar item. This is
different because \n in this string is converted to a space in a bar
item, so only \r actually creates a new line.

In my opinion, it's much more useful to know which line you focused on
in the bar item, and this is what the variable name sounds like it
represents. So the variable is changed to that.

Note that the coords->line variable is now unused. If it's useful, we
could assign it to a new variable in the focus hashtable, or if not we
should probably remove it.