core: fix bar_item_col for wrapped/new lines with horizontal filling
When the bar item filling is horizontal and the text of an item is
either wrapped or contains a \r so it's multiple lines and you focus on
a line that isn't the first, the bar_item_col value was incorrect. It
only accounted for focuses on the first line and set the col based on
where the item starts on that line, but for subsequent lines they start
at the beginning of the bar which is often a different column.