tools/check_curl_symbols.py: fix CURLOPT handling
The curl option handling seems to be non-functional since it suffers
from two distinct issues. The regular expression tries to match options
that we're not interested in and the symbol name is lacking the proper
prefix.

Adjust to match only on what we need and construct the name
appropriately... Fix the issues flagged by the updated script.

Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>